Refactir the formatting logic
* Modularized the formatting and linting components into logical files
* The formatter now runs the linter after formatting
* The linter now has the same progress feature the formatter has
* The linter now displays total errors as the final step before terminating
* The linters output is now a bit nicer and easier to read
