+++ package.json
@@ -21,6 +21,7 @@
   "license": "MIT",
   "dependencies": {
     "async": "^2.0.1",
+    "commander": "^2.9.0",
     "globby": "^6.0.0",
     "lodash.flatten": "^4.2.0",
     "lodash.range": "^3.1.5",
